revert BreakingSeedScore.Rank to int

+++ NET/Wdsf.Api.Client/Models/BreakingSeedScore.cs
@@ -13,6 +13,6 @@
 
         [XmlAttribute("rank")]
         [JsonProperty("rank")]
-        public decimal Rank { get; set; }
+        public int Rank { get; set; }
     }
 }
